change data-type

+++ db/migrate/20170725084925_create_erp_gift_givens_givens.rb
@@ -7,7 +7,7 @@ class CreateErpGiftGivensGivens < ActiveRecord::Migration[5.1]
       t.references :contact, index: true, references: :erp_contacts_contacts
       t.text :note
       t.string :status
-      t.decimal :cache_products_count
+      t.integer :cache_products_count
 
       t.timestamps
     end
